updated DeleteCommentService

+++ app/services/comments/delete_comment_service.rb
@@ -4,32 +4,25 @@ module Comments
   class DeleteCommentService
     include Callable
 
-    attr_reader :user_id, :column_id, :card_id, :id
+    attr_reader :current_user, :card_params, :id
 
-    def initialize(user_id, card_params, id)
-      @user_id = user_id
-      @column_id = card_params[:column_id]
-      @card_id = card_params[:card_id]
+    def initialize(current_user, card_params, id)
+      @card_params = {}
+
+      @current_user = current_user
+      @card_params[:column_id] = card_params[:column_id]
+      @card_params[:id] = card_params[:card_id]
       @id = id
     end
 
     def call
-      card = Card.find_by!(id: @card_id, column_id: @column_id)
+      raise Exceptions::Unauthorized, 'Unauthorized' unless @current_user.present?
+      card = Card.find_by!(@card_params)
       comment = card.comments.find_by!(id: @id)
-
+      raise Exceptions::Forbidden, 'Forbidden' unless comment.user == @current_user
       comment.destroy
-
-      if comment.destroyed?
-        result(true, {})
-      else
-        result(false, comment.errors)
-      end
-    end
-
-    private
-
-    def result(success, errors)
-      OpenStruct.new(success: success, errors: errors)
+      
+      OpenStruct.new(success: comment.destroyed?, errors: comment.errors)
     end
   end
 end
